api/requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
uvicorn[standard]==0.29.0
httpx==0.27.0
python-whois==0.9.4
dnspython==2.6.1
python-nmap==0.7.1
pydantic==2.7.1
